+++ http/requests/create_base64_image_input.go
@@ -3,4 +3,6 @@ package requests
 type CreateBase64ImageInput struct {
 	Base64    string `json:"base64" binding:"required"`
 	Extension string `json:"extension" binding:"required"`
+	Width     int    `json:"width"`
+	Height    int    `json:"height"`
 }